custom/requirements.txt
django-addanother
django-qr-code
